// src/types.ts
export type VacationRow = {
  BM: string;
  Nome: string;
  STATUS: string;
  "Divisão": string;
  "Data de início": string;
  "Data do término": string;
  "Dias Corrido": string;
  "Tipo de Férias": string;
  "Ultima altera": string;
  Subordinação: string;
  "RESPONSAVEL P/ LOTAÇÃO": string;
  "Nome do Próprio": string;

  _start?: Date | null;
  _end?: Date | null;
};